Move the connection pools initialiazitation logic to the pools module
